pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wavefinder"
version = "0.9.0"
authors = [
  { name="Cem Onyuksel", email="cem.onyuksel@cfa.harvard.edu" },
  { name="Colby Jergenson", email="colby.jurgenson@cfa.harvard.edu" }
]
description = "Measure optical performance at various wavelengths"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "astropy",
  "numpy",
  "Pillow",
  "pyserial",
  "pyusb",
  "libusb",
  "zaber-motion"
]
